// hub/frontend/lib/api/generated/model/skillFileRead.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Wardn Hub API
 * OpenAPI spec version: 0.2.95
 */
import type { SkillFileReadEncoding } from './skillFileReadEncoding';

export interface SkillFileRead {
  contents: string;
  encoding?: SkillFileReadEncoding;
  executable?: boolean;
  path: string;
}